validator on account router

+++ app/controllers/account.controller.js
@@ -6,28 +6,24 @@ const Op = db.Op;
 
 // Create and Save a new Account
 exports.create = async (req, res) => {
-  const errors = validationResult(req);
-  if (errors.isEmpty()) {
-    // user Id
-    const userId = req.userId;
+  // user Id
+  const userId = req.userId;
 
-    // Create an Account
-    const account = {
-      name: req.body.name,
-      balance: req.body.balance,
-      currencyTypeId: req.body.currencyTypeId,
-      userId: userId,
-    };
+  // Create an Account
+  const account = {
+    name: req.body.name,
+    balance: req.body.balance,
+    currencyTypeId: req.body.currencyTypeId,
+    userId: userId,
+  };
 
-    // Save Account in database
-    const result = await Account.create(account);
+  // Save Account in database
+  const result = await Account.create(account);
 
-    return res.status(200).json({
-      message: "Account created succesfully",
-      result: result,
-    });
-  }
-  res.status(422).json({ errors: errors.array() });
+  return res.status(200).json({
+    message: "Account created succesfully",
+    result: result,
+  });
 };
 
 // Retrieve all accounts from the database.
@@ -51,32 +47,66 @@ exports.findAll = (req, res) => {
           "Some error occurred while retrieving user's accounts.",
       });
     });
-
 };
 
-exports.delete = (req, res) => {
-  const id = req.params.id;
+exports.delete = async (req, res) => {
+  try {
+    // user Id
+    const userId = req.userId;
+    const id = req.params.id;
 
-  Account.destroy({
-    where: { id: id },
-  })
-    .then((num) => {
-      if (num == 1) {
-        res.send({
-          message: "Account was deleted successfully!",
-        });
-      } else {
-        res.send({
-          message: `Cannot delete Account with id=${id}. Maybe Account was not found!`,
-        });
-      }
-    })
-    .catch((err) => {
-      res.status(500).send({
-        message: "Could not delete Account with id=" + id,
-      });
+    // Find the account by id and userId
+    const account = await Account.findOne({
+      where: { id: id, userId: userId },
     });
+
+    if (!account) {
+      return res.status(404).send({
+        message: `Cannot delete Account with id=${id}. Account was not found or does not belong to the user!`,
+      });
+    }
+
+    // If account exists, delete it
+    await Account.destroy({
+      where: { id: id },
+    });
+
+    res.send({
+      message: "Account was deleted successfully!",
+    });
+  } catch (err) {
+    res.status(500).send({
+      message: "Could not delete Account with id=" + id,
+    });
+  }
 };
+
+
+// exports.delete = (req, res) => {
+//   // user Id
+//   const userId = req.userId;
+//   const id = req.params.id;
+
+//   Account.destroy({
+//     where: { id: id },
+//   })
+//     .then((num) => {
+//       if (num == 1) {
+//         res.send({
+//           message: "Account was deleted successfully!",
+//         });
+//       } else {
+//         res.send({
+//           message: `Cannot delete Account with id=${id}. Maybe Account was not found!`,
+//         });
+//       }
+//     })
+//     .catch((err) => {
+//       res.status(500).send({
+//         message: "Could not delete Account with id=" + id,
+//       });
+//     });
+// };
 
 exports.findOne = (req, res) => {
   const id = req.params.id;
@@ -128,7 +158,8 @@ exports.deleteAll = (req, res) => {
     })
     .catch((err) => {
       res.status(500).send({
-        message: err.message || "Some error occurred while removing all accounts.",
+        message:
+          err.message || "Some error occurred while removing all accounts.",
       });
     });
 };
